frontend/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
python-multipart==0.0.6